+++ CHANGE_LOG.md
@@ -2,12 +2,13 @@
 - Print summary string of what Wheatley is going to ring.
 - Add proper support for backstroke starts (with 3 rows of rounds for `--up-down-in`).
 - Allow support of the new tower sizes - `5`, `14` and `16`.
+- Add `-v`/`--verbose` and `-q`/`--quiet` to change how much stuff Wheatley prints.
 - Allow Wheatley to ring with any (positive) number of cover bells.
 - Add `--start-index` to specify how many rows into a lead of a method Wheatley should start.
 - Tell users when Wheatley is waiting for `Look To`.
 - Change place notation parsing to comply with CompLib and the XML specification.
 - Add full static typing, and fix some `None`-related bugs.
-- Prevent installing the wrong version of socketio to work with RingingRoom
+- Prevent installing the wrong version of socketio to work with RingingRoom.
 
 # 0.5.2
 - Bump numpy version to exactly `1.19.3` on Windows to fix